+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod console;
 pub mod file;
 pub mod http;